Adjusted the way to import yuanrong_client & Updated notes of tests
