+++ transfer_queue/storage/clients/__init__.py
@@ -13,3 +13,12 @@
 # limitations under the License.
 
 # This module is currently empty but reserved for future client implementations
+from .base import TransferQueueStorageKVClient
+from .factory import StorageClientFactory
+from .yuanrong_client import YRStorageClient
+
+__all__ = [
+    "TransferQueueStorageKVClient",
+    "StorageClientFactory",
+    "YRStorageClient",
+]
