+++ transfer_queue/storage/clients/factory.py
@@ -11,28 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from abc import ABC, abstractmethod
-
-from torch import Tensor
-
-
-class TransferQueueStorageClient(ABC):
-    """
-    Abstract base class for storage client.
-    Subclasses must implement the core methods: put, get, and clear.
-    """
-
-    @abstractmethod
-    def put(self, keys: list[str], values: list[Tensor]) -> None:
-        raise NotImplementedError("Subclasses must implement put")
-
-    @abstractmethod
-    def get(self, keys: list[str], shapes=None, dtypes=None) -> list[Tensor]:
-        raise NotImplementedError("Subclasses must implement get")
-
-    @abstractmethod
-    def clear(self, keys: list[str]) -> None:
-        raise NotImplementedError("Subclasses must implement clear")
+from transfer_queue.storage.clients.base import TransferQueueStorageKVClient
 
 
 class StorageClientFactory:
@@ -42,7 +21,7 @@ class StorageClientFactory:
     """
 
     # Class variable: maps client names to their corresponding classes
-    _registry: dict[str, TransferQueueStorageClient] = {}
+    _registry: dict[str, TransferQueueStorageKVClient] = {}
 
     @classmethod
     def register(cls, client_type: str):
@@ -54,14 +33,14 @@ class StorageClientFactory:
             Callable: The decorator function that returns the original class
         """
 
-        def decorator(client_class: TransferQueueStorageClient) -> TransferQueueStorageClient:
+        def decorator(client_class: TransferQueueStorageKVClient) -> TransferQueueStorageKVClient:
             cls._registry[client_type] = client_class
             return client_class
 
         return decorator
 
     @classmethod
-    def create(cls, client_type: str, config: dict) -> TransferQueueStorageClient:
+    def create(cls, client_type: str, config: dict) -> TransferQueueStorageKVClient:
         """
         Create and return an instance of the storage client by name.
         Args:
@@ -74,11 +53,3 @@ class StorageClientFactory:
         if client_type not in cls._registry:
             raise ValueError(f"Unknown StorageClient: {client_type}")
         return cls._registry[client_type](config)
-
-
-# TODO: Dynamically register the storage client class based on the configuration
-# Register storage clients
-try:
-    import transfer_queue.storage.clients.yuanrong_client as _  # noqa: F401
-except ImportError:
-    pass
